+++ cmd/hidden_lake/messenger/internal/handler/chat.go
@@ -8,18 +8,18 @@ import (
 
 	"github.com/number571/go-peer/cmd/hidden_lake/messenger/internal/app/state"
 	"github.com/number571/go-peer/cmd/hidden_lake/messenger/internal/database"
+	"github.com/number571/go-peer/cmd/hidden_lake/messenger/internal/utils"
+	pkg_settings "github.com/number571/go-peer/cmd/hidden_lake/messenger/pkg/settings"
 	"github.com/number571/go-peer/cmd/hidden_lake/messenger/web"
+	"github.com/number571/go-peer/cmd/hidden_lake/service/pkg/client"
 	"github.com/number571/go-peer/cmd/hidden_lake/service/pkg/request"
 	"github.com/number571/go-peer/internal/api"
+	"github.com/number571/go-peer/pkg/crypto/asymmetric"
 	"github.com/number571/go-peer/pkg/crypto/hashing"
 	"github.com/number571/go-peer/pkg/crypto/random"
+	"github.com/number571/go-peer/pkg/errors"
 
 	hlm_settings "github.com/number571/go-peer/cmd/hidden_lake/messenger/internal/settings"
-	pkg_settings "github.com/number571/go-peer/cmd/hidden_lake/messenger/pkg/settings"
-)
-
-const (
-	cChatLimitMessages = 1000
 )
 
 type sChatMessage struct {
@@ -62,51 +62,44 @@ func FriendsChatPage(pStateManager state.IStateManager) http.HandlerFunc {
 		}
 
 		client := pStateManager.GetClient().Service()
+
 		myPubKey, err := client.GetPubKey()
 		if err != nil {
 			fmt.Fprint(pW, "error: read public key")
 			return
 		}
 
-		friends, err := client.GetFriends()
+		recvPubKey, err := getReceiverPubKey(client, myPubKey, aliasName)
 		if err != nil {
-			fmt.Fprint(pW, "error: read friends")
+			fmt.Fprint(pW, err.Error())
 			return
 		}
 
-		friendPubKey, ok := friends[aliasName]
-		if !ok {
-			fmt.Fprint(pW, "undefined public key by alias name")
-			return
-		}
-
-		rel := database.NewRelation(myPubKey, friendPubKey)
+		rel := database.NewRelation(myPubKey, recvPubKey)
 		pR.ParseForm()
 
 		switch pR.FormValue("method") {
 		case http.MethodPost:
-			msg := strings.TrimSpace(pR.FormValue("input_message"))
+			rawMsg := strings.TrimSpace(pR.FormValue("input_message"))
+
+			// ReplaceTextToEmoji not used in trySendMessage
+			// because can be exists emoji (🕵️‍♂️ => 🕵️♂️) more than 4 bytes (rune)
+			// that creates more than 1 emoji after use GetOnlyWritableCharacters
+			msg := utils.GetOnlyWritableCharacters(rawMsg)
 			if msg == "" {
 				fmt.Fprint(pW, "error: message is null")
 				return
 			}
 
-			err := client.BroadcastRequest(
-				friendPubKey,
-				request.NewRequest(http.MethodPost, pkg_settings.CTitlePattern, hlm_settings.CPushPath).
-					WithHead(map[string]string{
-						"Content-Type": "application/json",
-					}).
-					WithBody([]byte(msg)),
-			)
-			if err != nil {
+			if err := trySendMessage(client, recvPubKey, myPubKey, msg); err != nil {
 				fmt.Fprint(pW, "error: push message to network")
 				return
 			}
 
 			uid := random.NewStdPRNG().GetBytes(hashing.CSHA256Size)
-			err = db.Push(rel, database.NewMessage(false, msg, uid))
-			if err != nil {
+			dbMsg := database.NewMessage(false, utils.ReplaceTextToEmoji(msg), uid)
+
+			if err := db.Push(rel, dbMsg); err != nil {
 				fmt.Fprint(pW, "error: add message to database")
 				return
 			}
@@ -119,8 +112,8 @@ func FriendsChatPage(pStateManager state.IStateManager) http.HandlerFunc {
 
 		start := uint64(0)
 		size := db.Size(rel)
-		if size > cChatLimitMessages {
-			start = size - cChatLimitMessages
+		if size > hlm_settings.CChatLimitMessages {
+			start = size - hlm_settings.CChatLimitMessages
 		}
 
 		msgs, err := db.Load(rel, start, size)
@@ -139,7 +132,7 @@ func FriendsChatPage(pStateManager state.IStateManager) http.HandlerFunc {
 			STemplateState: pStateManager.GetTemplate(),
 			FAddress: sChatAddress{
 				FClient: clientPubKey.GetAddress().ToString(),
-				FFriend: friendPubKey.GetAddress().ToString(),
+				FFriend: recvPubKey.GetAddress().ToString(),
 			},
 			FMessages: make([]sChatMessage, 0, len(msgs)),
 		}
@@ -162,4 +155,38 @@ func FriendsChatPage(pStateManager state.IStateManager) http.HandlerFunc {
 
 		t.Execute(pW, res)
 	}
+}
+
+func trySendMessage(client client.IClient, recvPubKey, myPubKey asymmetric.IPubKey, msg string) error {
+	// if the sender = receiver then there is no need to send a message to the network
+	if myPubKey.GetAddress().ToString() == recvPubKey.GetAddress().ToString() {
+		return nil
+	}
+
+	return client.BroadcastRequest(
+		recvPubKey,
+		request.NewRequest(http.MethodPost, pkg_settings.CTitlePattern, hlm_settings.CPushPath).
+			WithHead(map[string]string{
+				"Content-Type": "application/json",
+			}).
+			WithBody([]byte(msg)),
+	)
+}
+
+func getReceiverPubKey(client client.IClient, myPubKey asymmetric.IPubKey, aliasName string) (asymmetric.IPubKey, error) {
+	friends, err := client.GetFriends()
+	if err != nil {
+		return nil, errors.NewError("error: read friends")
+	}
+
+	if aliasName == hlm_settings.CIamAliasName {
+		return myPubKey, nil
+	}
+
+	friendPubKey, ok := friends[aliasName]
+	if !ok {
+		return nil, errors.NewError("undefined public key by alias name")
+	}
+
+	return friendPubKey, nil
 }